+++ GameStoreBack-end/WebApi/Controllers/GenreController.cs
@@ -4,6 +4,7 @@ using BLL.Services;
 using BLL.Interfaces;
 using BLL.Models;
 using WebApi.Models;
+using DLL.Entities;
 
 namespace WebApi.Controllers
 {
@@ -26,7 +27,7 @@ namespace WebApi.Controllers
             var genres = _mapper.Map<IEnumerable<GenreViewModel>>(await _genreService.GetAllAsync());
             return new JsonResult(genres);
         }
-        [HttpGet("{id}", Name = nameof(GetGenre))]
+        [HttpGet("{id}")]
         [ProducesResponseType(200)]
         [ProducesResponseType(404)]
         public async Task<IActionResult> GetGenre(int id)
@@ -42,22 +43,47 @@ namespace WebApi.Controllers
         [HttpPost]
         [ProducesResponseType(201)]
         [ProducesResponseType(400)]
-        public async Task<IActionResult> CreateGame(GenreViewModel game)
+        public async Task<IActionResult> CreateGenre(GenreViewModel genre)
         {
-            var genreModel = _mapper.Map<GenreModel>(game);
-            await _genreService.AddAsync(genreModel);
-            return Ok();
+            if (ModelState.IsValid)
+            {
+                var genreByName = await _genreService.GetByGenreNameAsync(genre.Name);
+                if(genreByName == null)
+                {
+                    return BadRequest();
+                }
+                var genreModel = _mapper.Map<GenreModel>(genre);
+                await _genreService.AddAsync(genreModel);
+                return Ok(genreModel);
+            }
+            else
+            {
+                return BadRequest();
+            }
         }
 
-        [HttpPut("{id}")]
+        [HttpPut]
         [ProducesResponseType(204)]
         [ProducesResponseType(400)]
         [ProducesResponseType(404)]
-        public async Task<IActionResult> Update(string id, [FromBody] GenreViewModel game)
+        public async Task<IActionResult> Update([FromBody] GenreViewModel genre)
         {
-            var genreModel = _mapper.Map<GenreModel>(game);
-            await _genreService.UpdateAsync(genreModel);
-            return NoContent();
+            if (ModelState.IsValid)
+            {
+                var gameByName = await _genreService.GetByGenreNameAsync(genre.Name);
+                if (gameById.Name != null && gameByName.Id != gameById.Id)
+                {
+                    return BadRequest();
+                }
+
+                var genreModel = _mapper.Map<GenreModel>(genre);
+                await _genreService.UpdateAsync(genreModel);
+                return Ok(genre);
+            }
+            else
+            {
+                return BadRequest();
+            }
         }
 
         [HttpDelete("{id}")]
@@ -66,8 +92,17 @@ namespace WebApi.Controllers
         [ProducesResponseType(404)]
         public async Task<IActionResult> Delete(int id)
         {
-            await _genreService.DeleteByIdAsync(id);
-            return NoContent();
+            var genre = await _genreService.GetByIdAsync(id);
+            if (genre == null)
+            {
+                return NotFound();
+            }
+            else
+            {
+                await _genreService.DeleteByIdAsync(id);
+                return Ok("Deleted");
+            }
+
         }
     }
 }
